requirements.txt
asgiref==3.5.2
asttokens==2.1.0
backcall==0.2.0
colorama==0.4.6
decorator==5.1.1
Django==3.0.14
django-rest-framework==0.1.0
djangorestframework==3.14.0
executing==1.2.0
ipdb==0.13.9
ipython==8.6.0
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.33
pure-eval==0.2.2
Pygments==2.13.0
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.1
toml==0.10.2
traitlets==5.5.0
wcwidth==0.2.5
